requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
openpyxl==3.1.2
python-multipart==0.0.6
Pillow==10.1.0
requests==2.31.0
opencv-python==4.8.1.78
numpy>=1.23.2,<2.0.0
